split qwen36 moe run policy helpers
